% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMBaug.R
\name{glmmTMBaug}
\alias{glmmTMBaug}
\title{Fit Penalized Generalized Linear Mixed Model via Data Augmentation}
\usage{
glmmTMBaug(
  formula,
  data,
  family,
  data_driven = TRUE,
  penOpt = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{formula}{A model formula specifying fixed and random effects, as in \code{glmmTMB}.}

\item{data}{A data frame containing the variables used in the model.}

\item{family}{family a family function, a character string naming a family function, or the result of a call to a family function (variance/link function) information. Only binomial(), poison() and gaussian() are currently supproted}

\item{data_driven}{Logical; if \code{TRUE}, estimates the penalty strength (\code{tau}) from the data.}

\item{penOpt}{A list of penalty options. Common elements include \code{tau}, \code{alpha}, \code{psi}, \code{nu}, \code{const}, \code{param}.}

\item{verbose}{Logical. If TRUE, prints details about the penalty used for fitting penalized model.}

\item{...}{Additional arguments passed to \code{glmmTMB()}.}
}
\value{
A list with elements:
\describe{
\item{non_pen}{The unpenalized \code{glmmTMB} model.}
\item{pen}{The penalized \code{glmmTMB} model fit with augmented data.}
\item{tau}{Estimated or supplied penalty strength.}
}
}
\description{
This function extends \code{glmmTMB} to fit generalized linear mixed models (GLMMs)
with a penalty on the random effects covariance matrix via a data augmentation approach.
}
\examples{
\dontrun{
data(sleepstudy, package = "lme4")
glmmTMB_aug(Reaction ~ Days + (Days|Subject), data = sleepstudy, family = gaussian())
}

}
