% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmmTMBaug.R
\name{glmmTMBaug}
\alias{glmmTMBaug}
\title{Fit Penalized Generalized Linear Mixed Model via Data Augmentation}
\usage{
glmmTMBaug(formula, data, family, penOpt = list(), verbose = TRUE, ...)
}
\arguments{
\item{formula}{A model formula specifying fixed and random effects, as in \code{glmmTMB}.
The current implementation supports penalization of a single grouping factor
(i.e., one level of random effects).
By default, the first random effect term specified in the formula is penalized.}

\item{data}{A data frame containing the variables used in the model.}

\item{family}{a family function, a character string naming a family function, or the result of a call to a family function (variance/link function) information. Only \code{binomial}, \code{poisson} and \code{gaussian} are currently supported.}

\item{penOpt}{A named list of penalty options used to control the penalized likelihood fit. If \code{psi=NULL} and \code{nu=NULL}, a data-driven approach is used to determine psi. If \code{tau=NULL} a data driven approach is used to determine tau.
If both \code{psi} and \code{nu} are specified, a prior with those parameters is used instead. In this case, the parameters \code{tau}, \code{trunc} and \code{alpha} are ignored.
The following penalty options are recognized (with defaults):

\itemize{
\item \code{tau} (default: \code{NULL}) — Numeric value in the interval [0, 1]. If \code{NULL}, a data-driven method is used to estimate it.
\item \code{trunc} (default: \code{c(1e-4, 1e4)}) — Lower and upper truncation bounds of eigenvalues in data-drive approach to determine \code{psi}.
\item \code{alpha} (default: \code{0.05}) — Significance level used in the data-driven procedure for estimating \code{tau}.
\item \code{psi} (default: \code{NULL}) — A positive-definite matrix specifying the prior scale matrix for the inverse Wishart (variance) or Wishart (precision) distribution. Required if using a fixed predefined penalty.
\item \code{nu} (default: \code{NULL}) — Degrees of freedom for the prior distribution. Required if using a fixed predefined penalty. Must be an integer and compatible with the chosen \code{param}. For example, \code{(nu + q + 1)/q} must be an integer if \code{param = "variance"} and \code{(nu - q - 1)/q} must be an integer if \code{param = "precision"}.
\item \code{const} (default: \code{1e8}) — A constant used in data augmentation to set the value of precision weights for pseudo-observations. Larger values may provide better approximation of the penalty, but may introduce numerical instability. Small values may result in poor approximation of the penalty.
\item \code{param} (default: \code{"variance"}) — Specifies the parametrization of the prior: either \code{"variance"} for an inverse Wishart prior on the variance-covariance matrix or \code{"precision"} for a Wishart prior on the precision matrix. For univariate random effect, \code{"log variance"} can also be specified.
}}

\item{verbose}{Logical. If TRUE, prints details about the penalty used for fitting penalized model and messages about the penalized model fitting.}

\item{...}{Additional arguments passed to \code{glmmTMB()}.}
}
\value{
The function returns a list with elements:

\describe{
\item{non_pen}{The non-penalized \code{glmmTMB} model.}
\item{pen}{The penalized \code{glmmTMB} model fit with augmented data.}
\item{tau}{Estimated or supplied shrinkage strength towards mean eigenvalue. If the penalty was prespecified the value is \code{NULL}.}
\item{error_pen}{If penalized fitting fails, the function returns the original unpenalized glmmTMB fit with pen = NULL and a flag error_pen = TRUE to indicate the failure.}
}
}
\description{
This function extends \code{glmmTMB} to fit generalized linear mixed models (GLMMs)
with a penalty on the random effects covariance matrix via a data augmentation approach.
}
\details{
When the random effect structure is univariate, the prior distribution on the variance is Inverse Gamma distribution that equals to univariate Inverse Wishart distribution.
To implement an Inverse Gamma penalty (or some other parametrization; see \code{penOpt} parameter \code{param} for details) with specified shape (\eqn{\alpha}) and scale (\eqn{\beta}),
set \code{nu=}\eqn{2\alpha} and \code{psi= }\eqn{2\beta}.

Augmented approach in Poisson model with log link is implemented by the offset approach to improve numerical stability. Poisson models with different link functions
are implemented using weights. For these models the recommended value is \code{cons=10^6}.
}
\examples{
\dontrun{

library(blmeco)
data(ellenberg)
ellenberg$gradient  <-  paste(ellenberg$Year, ellenberg$Soil)
ellenberg$water.z <- as.numeric(scale(ellenberg$Water))
glmmTMBaug(log(Yi.g) ~ water.z + I(water.z^2) +
          (water.z + I(water.z^2)|Species) + (1|gradient), data=ellenberg, family="gaussian")

}
}
\references{
To be added after the article is published.
}
